sql/catalog/accessor: extract the logical and physical accessors to a package
Release note: None
