+++ pkg/sql/catalog/accessors/logical_schema_accessors.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package sql
+package accessors
 
 import (
 	"context"
@@ -17,7 +17,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/errors"
@@ -27,13 +26,24 @@ import (
 // interfaces defined in schema_accessors.go.
 //
 
+// NewLogicalAccessor constructs a new accessor given an underlying physical
+// accessor and VirtualSchemas.
+func NewLogicalAccessor(
+	physicalAccessor catalog.Accessor, vs catalog.VirtualSchemas,
+) *LogicalSchemaAccessor {
+	return &LogicalSchemaAccessor{
+		Accessor: physicalAccessor,
+		vs:       vs,
+	}
+}
+
 // LogicalSchemaAccessor extends an existing DatabaseLister with the
 // ability to list tables in a virtual schema.
 type LogicalSchemaAccessor struct {
 	catalog.Accessor
 	vs catalog.VirtualSchemas
 	// Used to avoid allocations.
-	tn TableName
+	tn tree.TableName
 }
 
 var _ catalog.Accessor = &LogicalSchemaAccessor{}
@@ -55,10 +65,10 @@ func (l *LogicalSchemaAccessor) GetObjectNames(
 	ctx context.Context,
 	txn *kv.Txn,
 	codec keys.SQLCodec,
-	dbDesc *DatabaseDescriptor,
+	dbDesc *sqlbase.DatabaseDescriptor,
 	scName string,
 	flags tree.DatabaseListFlags,
-) (TableNames, error) {
+) (tree.TableNames, error) {
 
 	if entry, ok := l.vs.GetVirtualSchema(scName); ok {
 		names := make(tree.TableNames, 0, entry.NumTables())
@@ -90,8 +100,9 @@ func (l *LogicalSchemaAccessor) GetObjectDesc(
 	case tree.TypeObject:
 		// TODO(ajwerner): Change this function if we ever expose non-table objects
 		// underneath virtual schemas. For now we assume that the only objects
-		// ever handed back from GetObjectByName is
-		return (catalogkv.UncachedPhysicalAccessor{}).GetObjectDesc(ctx, txn, settings, codec, db, schema, object, flags)
+		// ever handed back from GetObjectByName are tables. Instead we fallthrough
+		// to the underlying physical accessor.
+		return l.Accessor.GetObjectDesc(ctx, txn, settings, codec, db, schema, object, flags)
 	case tree.TableObject:
 		l.tn = tree.MakeTableNameWithSchema(tree.Name(db), tree.Name(schema), tree.Name(object))
 		if scEntry, ok := l.vs.GetVirtualSchema(schema); ok {
